# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UstaadAssist API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Backend"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The `access_token` Supabase Auth returns when the teacher signs in, inside the app. Send it as `Authorization: Bearer <token>`. This API only verifies it — it never issues one."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string",
            "example": "Course not found"
          }
        },
        "required": [
          "success",
          "data",
          "message"
        ]
      },
...
      "CreateCourseBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Database Systems"
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "example": "CS-301"
          },
          "semester": {
            "type": [
              "string",
              "null"
            ],
            "example": "Fall 2026"
          },
          "start_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2026-09-01",
            "description": "A calendar date, YYYY-MM-DD"
          },
          "end_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2026-09-01",
            "description": "A calendar date, YYYY-MM-DD"
          },
          "class_days": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "mon",
                "tue",
                "wed",
                "thu",
                "fri",
                "sat",
                "sun"
              ]
            },
            "minItems": 1,
            "maxItems": 7,
            "example": [
              "mon",
              "wed"
            ],
            "description": "The days of the week class is held"
          }
        },
        "required": [
          "name",
          "start_date",
          "end_date",
          "class_days"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/courses": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Create a course",
        "description": "The minimum to get started: name, dates and class days. Public holidays for the date range are preloaded automatically in the same transaction. Students, weightage and the grade scale are asked for later, by the features that need them.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a course",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Course"
                    },
                    "message": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Something in the request was wrong — a bad value, a missing field, or a body that is not valid JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A bug on our side. The real error is in the server log, never in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "List the teacher's courses",
        "description": "Includes counts and a has_plan flag, so the list screen needs no extra requests.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List the teacher's courses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CourseListItem"
                      }
                    },
                    "message": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A bug on our side. The real error is in the server log, never in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{courseId}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "One course, with its topics and holidays",
        "description": "Everything the setup screen needs in one request.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "8"
            },
            "required": true,
            "name": "courseId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "One course, with its topics and holidays",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CourseDetail"
                    },
                    "message": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Does not exist — or belongs to another teacher. Both answer 404 on purpose: a 403 would confirm the id is real.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A bug on our side. The real error is in the server log, never in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Courses"
        ],
        "summary": "Edit course settings",
        "description": "Only the fields you send are changed. Changing the dates or class days invalidates the timetable, so the response carries replan_required — show the teacher a 'rebuild your plan' prompt when it is true. Extending the semester also tops up any newly exposed public holidays, leaving the teacher's existing ticks alone.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "8"
            },
            "required": true,
            "name": "courseId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Edit course settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Course"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "replan_required": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "replan_required"
                          ]
                        }
                      ]
                    },
                    "message": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Something in the request was wrong — a bad value, a missing field, or a body that is not valid JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Does not exist — or belongs to another teacher. Both answer 404 on purpose: a 403 would confirm the id is real.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A bug on our side. The real error is in the server log, never in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{courseId}/clone": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Start next semester from this course",
        "description": "Copies the teacher's preparation: topics (reset to pending), weightage and grade scale. Does NOT copy students, sessions, attendance or marks — those belong to the semester that ended. Public holidays are recalculated for the new dates rather than copied, because Eid moves every year.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "8"
            },
            "required": true,
            "name": "courseId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneCourseBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Start next semester from this course",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Course"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "copied": {
                              "type": "object",
                              "properties": {
                                "topics": {
                                  "type": "integer"
                                },
                                "holidays": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "topics",
                                "holidays"
                              ]
                            }
                          },
                          "required": [
                            "copied"
                          ]
                        }
                      ]
                    },
                    "message": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Something in the request was wrong — a bad value, a missing field, or a body that is not valid JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Does not exist — or belongs to another teacher. Both answer 404 on purpose: a 403 would confirm the id is real.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A bug on our side. The real error is in the server log, never in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{courseId}/topics": {
      "post": {
        "tags": [
          "Topics"
        ],
        "summary": "Set the course topics from one block of text",
        "description": "Send the whole textarea as `raw`, one topic per line. The server splits, trims and drops blank lines. Only the title is needed — sessions_needed and min_sessions default to 1 and priority to normal, and the teacher corrects them after seeing the generated plan. NOTE: this REPLACES all topics for the course.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "8"
            },
            "required": true,
            "name": "courseId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTopicsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Set the course topics from one block of text",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Topic"
                      }
                    },
                    "message": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "message"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Something in the request was wrong — a bad value, a missing field, or a body that is not valid JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No token, or a token that does not verify.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Does not exist — or belongs to another teacher. Both answer 404 on purpose: a 403 would confirm the id is real.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A bug on our side. The real error is in the server log, never in the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}